requirements.txt
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.2
django-extensions==3.2.3
djangorestframework==3.14.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
pillow==10.2.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
whitenoise==6.6.0
zope.interface==6.2
